Safe protection of runtime out_of_range error from the map
The UrQMD output files sometimes generate an entry with all zeros for
particle information. Such an entry will trigger the out_of_range error
when convert the urqmd_id to the pdg id using c++ map. Now, we implement
the try and catch protection to print an error message on screen when
such invalid entry is encountered. We will simply ignore these entries
and continue the analysis without them.