Boost particles with a global rap_shift
In asymmetric collision system, the measurements were usually not
presented in the center of mass frame but lab frame. So one needs to
perform a global boost for all particle.

In the meantime, we define rapidity and pseudo-rapidity in the
particle_info to be avoid confusing convention determined by rap_type
during the analysis.

We use local variable paraRdr when we redefine the rapidity and pT cuts
in the Analysis class so that the paraRdr_ is not touched for other
analysis.