update the anaylysis script and add parameters to ths list
We change the rapidity range for STAR flow from [-1, -0.1] and [0.1, 1]
to [-1, -0.5] and [0.5, 1]. In this way, we can use these two files to
compute the flow with an eta gap of 1, which are far more practical than
the original one with a eta gap of 0.2.

If a parameter is not found, a default value will be used. In this case,
we will add this parameter to the list with its default value. In this
way, we will not see repeat warnings in the output.